requirements.txt
streamlit
python-dotenv
langchain
langchain-core
langchain-community
langchain-google-genai
langchain-chroma
chromadb
pypdf
docx2txt
unstructured
python-pptx
duckduckgo-search
fpdf2
Pillow